Use double precision for touch event valuator
Touch event's "Touch Timestamp" valuator loses precision
if stored in float. So we change ExtractValuator() to store
the extracted valuator in double type.

BUG=None.
TEST=Make sure the touch timestamp from Chrome side is exactly
     the same as they are set in kernel evdev.

Review URL: https://chromiumcodereview.appspot.com/15792002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201890 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a76456e47738cf47ee3a20cbd749bfed53976931

+++ ash/touch/touch_observer_hud.cc
@@ -89,7 +89,7 @@ int GetTrackingId(const ui::TouchEvent& event) {
     return 0;
 #if defined(USE_XI2_MT)
   ui::ValuatorTracker* valuators = ui::ValuatorTracker::GetInstance();
-  float tracking_id;
+  double tracking_id;
   if (valuators->ExtractValuator(*event.native_event(),
                                  ui::ValuatorTracker::VAL_TRACKING_ID,
                                  &tracking_id)) {
